refs #29574 code review ve delete + properties

+++ tekir-recruit/src/main/java/com/ozguryazilim/tekir/recruit/jobapplication/information/AbstractJobApplicationInformationEditor.java
@@ -110,8 +110,12 @@ public abstract class AbstractJobApplicationInformationEditor<E extends EntityBa
 
     @Transactional
     public void delete(E entity) {
-        getRepository().remove(entity);
+        if (isJobApplicationViewPage()) {
+            getRepository().remove(entity);
+        }
+
         getEntityList().remove(entity);
+        RequestContext.getCurrentInstance().closeDialog(null);
     }
 
     @Transactional
